ready to push, may need cleaning
