--- python/sglang/srt/models/granitemoeshared.py
@@ -1,404 +1,0 @@
-"""Inference-only GraniteMoeShared model.
-
-The architecture is the same as granitemoe but with the addition of shared
-experts.
-"""
-
-from typing import Iterable, List, Optional, Tuple, Union
-
-import torch
-from torch import nn
-from transformers.models.granitemoeshared import GraniteMoeSharedConfig
-
-from sglang.srt.distributed import get_pp_group
-from sglang.srt.layers.activation import SiluAndMul
-from sglang.srt.layers.layernorm import RMSNorm
-from sglang.srt.layers.linear import MergedColumnParallelLinear, RowParallelLinear
-from sglang.srt.layers.logits_processor import LogitsProcessor
-from sglang.srt.layers.pooler import Pooler, PoolingType
-from sglang.srt.layers.quantization.base_config import QuantizationConfig
-from sglang.srt.layers.utils import PPMissingLayer
-from sglang.srt.layers.vocab_parallel_embedding import (
-    ParallelLMHead,
-    VocabParallelEmbedding,
-)
-from sglang.srt.model_executor.forward_batch_info import ForwardBatch, PPProxyTensors
-from sglang.srt.models.transformers import maybe_prefix
-from sglang.srt.server_args import get_global_server_args
-from sglang.srt.utils import add_prefix, make_layers
-
-from .granitemoe import GraniteMoeAttention, GraniteMoeModel, GraniteMoeMoE
-
-
-class GraniteMoeSharedMLP(nn.Module):
-    def __init__(
-        self,
-        config: GraniteMoeSharedConfig,
-        quant_config: QuantizationConfig | None = None,
-        prefix: str = "",
-    ):
-        super().__init__()
-
-        self.input_size = config.hidden_size
-        self.hidden_size = config.shared_intermediate_size
-        self.input_linear = MergedColumnParallelLinear(
-            input_size=self.input_size,
-            output_sizes=[self.hidden_size] * 2,
-            bias=False,
-            quant_config=quant_config,
-            prefix=f"{prefix}.input_linear",
-        )
-        self.output_linear = RowParallelLinear(
-            self.hidden_size,
-            self.input_size,
-            bias=False,
-            quant_config=quant_config,
-            prefix=f"{prefix}.output_linear",
-        )
-        if config.hidden_act != "silu":
-            raise ValueError(
-                f"Unsupported activation: {config.hidden_act}. "
-                "Only silu is supported for now."
-            )
-        self.act_fn = SiluAndMul()
-
-    def forward(self, hidden_states: torch.Tensor) -> torch.Tensor:
-        gate_up, _ = self.input_linear(hidden_states)
-        x = self.act_fn(gate_up)
-        x, _ = self.output_linear(x)
-        return x
-
-
-class GraniteMoeSharedDecoderLayer(nn.Module):
-    def __init__(
-        self,
-        config: GraniteMoeSharedConfig,
-        layer_id: int = 0,
-        quant_config: QuantizationConfig | None = None,
-        prefix: str = "",
-    ) -> None:
-        super().__init__()
-        self.hidden_size = config.hidden_size
-        self.self_attn = GraniteMoeAttention(
-            hidden_size=self.hidden_size,
-            num_heads=config.num_attention_heads,
-            num_kv_heads=config.num_key_value_heads,
-            max_position=config.max_position_embeddings,
-            layer_id=layer_id,
-            rope_theta=config.rope_theta,
-            quant_config=quant_config,
-            attention_multiplier=config.attention_multiplier,
-            prefix=f"{prefix}.self_attn",
-        )
-        self.block_sparse_moe = GraniteMoeMoE(
-            num_experts=config.num_local_experts,
-            top_k=config.num_experts_per_tok,
-            hidden_size=config.hidden_size,
-            intermediate_size=config.intermediate_size,
-            layer_id=layer_id,
-            quant_config=quant_config,
-            prefix=f"{prefix}.block_sparse_moe",
-        )
-        self.shared_mlp = (
-            None
-            if getattr(config, "shared_intermediate_size", 0) == 0
-            else GraniteMoeSharedMLP(
-                config, quant_config=quant_config, prefix=f"{prefix}.shared_mlp"
-            )
-        )
-
-        self.input_layernorm = RMSNorm(config.hidden_size, eps=config.rms_norm_eps)
-        self.post_attention_layernorm = RMSNorm(
-            config.hidden_size, eps=config.rms_norm_eps
-        )
-
-        self.residual_multiplier = config.residual_multiplier
-
-    def forward(
-        self,
-        positions: torch.Tensor,
-        hidden_states: torch.Tensor,
-    ) -> torch.Tensor:
-        # Self Attention
-        residual = hidden_states
-        hidden_states = self.input_layernorm(hidden_states)
-        hidden_states = self.self_attn(
-            positions=positions,
-            hidden_states=hidden_states,
-        )
-        hidden_states = residual + hidden_states * self.residual_multiplier
-        residual = hidden_states
-        hidden_states = self.post_attention_layernorm(hidden_states)
-        if self.shared_mlp is None:
-            hidden_states = self.block_sparse_moe(hidden_states)
-        else:
-            # create a copy since block_sparse_moe modifies in-place
-            moe_hidden_states = hidden_states.clone()
-            moe_hidden_states = self.block_sparse_moe(moe_hidden_states)
-            hidden_states = moe_hidden_states + self.shared_mlp(hidden_states)
-            del moe_hidden_states
-        hidden_states = residual + hidden_states * self.residual_multiplier
-
-        return hidden_states
-
-
-class GraniteMoeSharedModel(nn.Module):
-    def __init__(
-        self,
-        config: GraniteMoeSharedConfig,
-        quant_config: Optional[QuantizationConfig] = None,
-        prefix: str = "",
-    ):
-        super().__init__()
-
-        self.config = config
-        self.quant_config = quant_config  # Required by MixtralModel
-        self.padding_idx = config.pad_token_id
-        self.vocab_size = config.vocab_size
-
-        self.pp_group = get_pp_group()
-
-        if self.pp_group.is_first_rank:
-            self.embed_tokens = VocabParallelEmbedding(
-                config.vocab_size,
-                config.hidden_size,
-                quant_config=quant_config,
-                prefix=add_prefix("embed_tokens", prefix),
-            )
-        else:
-            self.embed_tokens = PPMissingLayer()
-
-        self.embedding_multiplier = config.embedding_multiplier
-
-        self.start_layer, self.end_layer, self.layers = make_layers(
-            config.num_hidden_layers,
-            lambda idx, prefix: GraniteMoeSharedDecoderLayer(
-                config, layer_id=idx, quant_config=self.quant_config, prefix=prefix
-            ),
-            prefix=f"{prefix}.layers",
-        )
-
-        if self.pp_group.is_last_rank:
-            self.norm = RMSNorm(config.hidden_size, eps=config.rms_norm_eps)
-        else:
-            self.norm = PPMissingLayer(return_tuple=True)
-
-        self.layers_to_capture = []
-
-    def get_input_embeddings(self) -> nn.Embedding:
-        """Get input embeddings from the model."""
-        return self.embed_tokens
-
-    def forward(
-        self,
-        input_ids: torch.Tensor | None,
-        positions: torch.Tensor,
-        forward_batch: ForwardBatch,
-        inputs_embeds: torch.Tensor | None = None,
-        pp_proxy_tensors: Optional[PPProxyTensors] = None,
-    ) -> Union[torch.Tensor, Tuple[torch.Tensor, List[torch.Tensor]], PPProxyTensors]:
-
-        if self.pp_group.is_first_rank:
-            if inputs_embeds is not None:
-                hidden_states = inputs_embeds
-            else:
-                hidden_states = self.embed_tokens(input_ids)
-            hidden_states *= self.embedding_multiplier
-
-        else:
-            assert pp_proxy_tensors is not None
-            hidden_states = pp_proxy_tensors.pp_proxy_tensors["hidden_states"]
-            residual = pp_proxy_tensors.pp_proxy_tensors["residual"]
-
-        aux_hidden_states = []
-        for i in range(self.start_layer, self.end_layer):
-            if i in self.layers_to_capture:
-                aux_hidden_states.append(hidden_states + residual)
-            layer = self.layers[i]
-            hidden_states, residual = layer(
-                positions,
-                hidden_states,
-                forward_batch,
-                residual,
-            )
-
-        if not self.pp_group.is_last_rank:
-            return PPProxyTensors(
-                {
-                    "hidden_states": hidden_states,
-                    "residual": residual,
-                }
-            )
-        else:
-            hidden_states, _ = self.norm(hidden_states, residual)
-
-        if len(aux_hidden_states) == 0:
-            return hidden_states
-
-        return hidden_states, aux_hidden_states
-
-
-class GraniteMoeSharedForCausalLM(nn.Module):
-    fall_back_to_pt_during_load = False
-
-    packed_modules_mapping = {
-        "qkv_proj": [
-            "q_proj",
-            "k_proj",
-            "v_proj",
-        ],
-    }
-
-    # LoRA specific attributes
-    embedding_modules = {
-        "embed_tokens": "input_embeddings",
-        "lm_head": "output_embeddings",
-    }
-
-    def __init__(
-        self,
-        config: GraniteMoeSharedConfig,
-        quant_config: Optional[QuantizationConfig] = None,
-        prefix: str = "",
-    ):
-        super().__init__()
-
-        self.config = config
-        self.pp_group = get_pp_group()
-        self.quant_config = quant_config
-
-        if self.config.tie_word_embeddings:
-            self.lm_head = self.model.embed_tokens
-        else:
-            self.lm_head = ParallelLMHead(
-                config.vocab_size,
-                config.hidden_size,
-                quant_config=quant_config,
-                prefix=add_prefix("lm_head", prefix),
-                use_attn_tp_group=get_global_server_args().enable_dp_lm_head,
-            )
-
-        self.logits_processor = LogitsProcessor(
-            config, logit_scale=1 / self.config.logits_scaling
-        )
-        self.pooler = Pooler(pooling_type=PoolingType.LAST, normalize=True)
-
-        self.model = GraniteMoeSharedModel(
-            self.config, quant_config=quant_config, prefix=maybe_prefix(prefix, "model")
-        )
-        self.capture_aux_hidden_states = False
-
-    def embed_input_ids(self, input_ids: torch.Tensor) -> torch.Tensor:
-        return self.model.embed_input_ids(input_ids)
-
-    def forward(
-        self,
-        input_ids: torch.Tensor,
-        positions: torch.Tensor,
-        forward_batch: ForwardBatch,
-        input_embeds: torch.Tensor = None,
-        get_embedding: bool = False,
-        pp_proxy_tensors: Optional[PPProxyTensors] = None,
-    ) -> torch.Tensor:
-        hidden_states = self.model(
-            input_ids=input_ids,
-            positions=positions,
-            forward_batch=forward_batch,
-            inputs_embeds=input_embeds,
-            pp_proxy_tensors=pp_proxy_tensors,
-        )
-
-        aux_hidden_states = None
-        if self.capture_aux_hidden_states:
-            hidden_states, aux_hidden_states = hidden_states
-
-        if self.pp_group.is_last_rank:
-            if not get_embedding:
-                return self.logits_processor(
-                    input_ids,
-                    hidden_states,
-                    self.lm_head,
-                    forward_batch,
-                    aux_hidden_states,
-                )
-            else:
-                return self.pooler(hidden_states, forward_batch)
-        else:
-            return hidden_states
-
-    def compute_logits(self, hidden_states: torch.Tensor) -> torch.Tensor | None:
-        logits = self.logits_processor(self.lm_head, hidden_states)
-        return logits
-
-    def get_input_embeddings(self) -> nn.Embedding:
-        return self.model.embed_tokens
-
-    def get_embed_and_head(self):
-        return self.model.embed_tokens.weight, self.lm_head.weight
-
-    def set_embed_and_head(self, embed, head):
-        del self.model.embed_tokens.weight
-        del self.lm_head.weight
-        self.model.embed_tokens.weight = embed
-        self.lm_head.weight = head
-        torch.cuda.empty_cache()
-        torch.cuda.synchronize()
-
-    def get_embed(self):
-        return self.model.embed_tokens.weight
-
-    def load_weights(self, weights: Iterable[tuple[str, torch.Tensor]]):
-        new_weights = {}
-        for n, p in weights:
-            if n.endswith(".block_sparse_moe.input_linear.weight"):
-                for e in range(p.size(0)):
-                    w1_name = n.replace(
-                        ".block_sparse_moe.input_linear.weight",
-                        f".block_sparse_moe.experts.{e}.w1.weight",
-                    )
-                    w3_name = n.replace(
-                        ".block_sparse_moe.input_linear.weight",
-                        f".block_sparse_moe.experts.{e}.w3.weight",
-                    )
-                    w1_param, w3_param = p[e].chunk(2, dim=0)
-                    assert w1_name not in new_weights
-                    assert w3_name not in new_weights
-                    new_weights[w1_name] = w1_param
-                    new_weights[w3_name] = w3_param
-            elif n.endswith(".block_sparse_moe.output_linear.weight"):
-                for e in range(p.size(0)):
-                    w2_name = n.replace(
-                        ".block_sparse_moe.output_linear.weight",
-                        f".block_sparse_moe.experts.{e}.w2.weight",
-                    )
-                    w2_param = p[e]
-                    assert w2_name not in new_weights
-                    new_weights[w2_name] = w2_param
-            elif n.endswith(".block_sparse_moe.router.layer.weight"):
-                gate_name = n.replace(
-                    ".block_sparse_moe.router.layer.weight",
-                    ".block_sparse_moe.gate.weight",
-                )
-                assert gate_name not in new_weights
-                new_weights[gate_name] = p
-            else:
-                new_weights[n] = p
-        GraniteMoeModel._load_weights(self, new_weights.items())
-
-    # def make_empty_intermediate_tensors(
-    #     self, batch_size: int, dtype: torch.dtype, device: torch.device
-    # ) -> IntermediateTensors:
-    #     return IntermediateTensors(
-    #         {
-    #             "hidden_states": torch.zeros(
-    #                 (batch_size, self.config.hidden_size), dtype=dtype, device=device
-    #             ),
-    #         }
-    #     )
-
-    # def load_weights(self, weights: Iterable[tuple[str, torch.Tensor]]) -> set[str]:
-    #     loader = AutoWeightsLoader(
-    #         self,
-    #         skip_prefixes=(["lm_head."] if self.config.tie_word_embeddings else None),
-    #     )
-    #     return loader.load_weights(weights)
